Move the traversal traits into the traversal module.
